% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregate_raster}
\alias{aggregate_raster}
\title{Aggregate raster with proper handling of circular variables}
\usage{
aggregate_raster(rast, fact, circular = FALSE, fun = "mean")
}
\arguments{
\item{rast}{SpatRaster object}

\item{fact}{Integer, aggregation factor}

\item{circular}{Logical, whether variable is circular (e.g., aspect)}

\item{fun}{Character, aggregation function ("mean", "sum", "modal")}
}
\value{
Aggregated SpatRaster
}
\description{
Aggregate raster with proper handling of circular variables
}
\keyword{internal}
