% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_aggregation_factor}
\alias{calc_aggregation_factor}
\title{Calculate aggregation factor between two resolutions}
\usage{
calc_aggregation_factor(native_res, target_res)
}
\arguments{
\item{native_res}{Numeric, native resolution in meters}

\item{target_res}{Numeric, target resolution in meters}
}
\value{
Integer aggregation factor
}
\description{
Calculate aggregation factor between two resolutions
}
\keyword{internal}
