% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_tile_names}
\alias{calculate_tile_names}
\title{Generate tile names for 10x10 degree global grid}
\usage{
calculate_tile_names(bbox, tile_size = 10)
}
\arguments{
\item{bbox}{Numeric vector of length 4 (xmin, ymin, xmax, ymax)}

\item{tile_size}{Numeric, size of tiles in degrees (default 10)}
}
\value{
Character vector of tile names in format "NNx_EEWW"
}
\description{
Generate tile names for 10x10 degree global grid
}
\keyword{internal}
