% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifl.R
\name{download_ifl}
\alias{download_ifl}
\title{Download Intact Forest Landscapes shapefile}
\usage{
download_ifl(output_folder = "data/IFL", timeout = 600, year = 2016)
}
\arguments{
\item{output_folder}{Character, directory to save and extract files
(default: "data/IFL")}

\item{timeout}{Numeric, download timeout in seconds (default: 600)}

\item{year}{Numeric, year to download (2000, 2013, 2016, or 2020). Default: 2016}
}
\value{
Character, path to the extracted shapefile
}
\description{
Downloads IFL 2016 shapefile from the official Intact Forests website.
}
\keyword{internal}
