% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensure_directory}
\alias{ensure_directory}
\title{Create output directory if it doesn't exist}
\usage{
ensure_directory(dir_path)
}
\arguments{
\item{dir_path}{Character, directory path}
}
\value{
NULL (creates directory as side effect)
}
\description{
Create output directory if it doesn't exist
}
\keyword{internal}
