% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esacci_biomass.R
\name{esacci_tile_names}
\alias{esacci_tile_names}
\title{Generate ESA CCI AGB tile names for a region}
\usage{
esacci_tile_names(roi, year, version, type = "agb")
}
\arguments{
\item{roi}{sf object, SpatVector, or numeric bbox}

\item{year}{Numeric, year}

\item{version}{Character, version string}

\item{type}{Character, either "agb" or "sd" for standard deviation}
}
\value{
Character vector of tile filenames
}
\description{
Generate ESA CCI AGB tile names for a region
}
\keyword{internal}
