% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esacci_biomass.R
\name{ESACCIAGBtileNames}
\alias{ESACCIAGBtileNames}
\title{Generate ESA-CCI AGB tile names (Plot2Map compatible)}
\usage{
ESACCIAGBtileNames(
  pol,
  esacci_biomass_year = "latest",
  esacci_biomass_version = "latest"
)
}
\arguments{
\item{pol}{An sf or SpatVector object representing the polygon of interest.}

\item{esacci_biomass_year}{Numeric or "latest", year to download (2007, 2010, 2015-2022)}

\item{esacci_biomass_version}{Character or "latest", version to download (v2.0-v6.0)}
}
\value{
A character vector of unique file names for ESA-CCI AGB tiles.
}
\description{
This function generates file names for ESA-CCI AGB tiles based on a given polygon.
}
