% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glad_tcc.R
\name{glad_tcc_tile_names}
\alias{glad_tcc_tile_names}
\title{Generate GLAD TCC 2010 tile names for a region}
\usage{
glad_tcc_tile_names(roi)
}
\arguments{
\item{roi}{sf object, SpatVector, or numeric bbox}
}
\value{
Character vector of tile names in format "20N_110W"
}
\description{
Generate GLAD TCC 2010 tile names for a region
}
\keyword{internal}
