% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansen_gfc.R
\name{hansen_tile_names}
\alias{hansen_tile_names}
\title{Generate Hansen GFC tile names for a region}
\usage{
hansen_tile_names(roi)
}
\arguments{
\item{roi}{sf object, SpatVector, or numeric bbox}
}
\value{
Character vector of tile names in format "10N_080W"
}
\description{
Generate Hansen GFC tile names for a region
}
\keyword{internal}
