% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_tiles_for_roi}
\alias{list_tiles_for_roi}
\title{List tiles that intersect with a region of interest}
\usage{
list_tiles_for_roi(roi, tiles_dir, pattern)
}
\arguments{
\item{roi}{sf object or bbox representing region of interest}

\item{tiles_dir}{Character, directory containing tile files}

\item{pattern}{Character, regex pattern to match tile files}
}
\value{
Character vector of file paths
}
\description{
List tiles that intersect with a region of interest
}
\keyword{internal}
