% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potapov_height.R
\name{potapov_tile_names}
\alias{potapov_tile_names}
\title{Generate Potapov height tile names for a region}
\usage{
potapov_tile_names(roi)
}
\arguments{
\item{roi}{sf object, SpatVector, or numeric bbox}
}
\value{
Character vector of tile names
}
\description{
Generate Potapov height tile names for a region
}
\keyword{internal}
