% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sexton_treecover.R
\name{sexton_tile_names}
\alias{sexton_tile_names}
\title{Generate Sexton tree cover tile names for a region}
\usage{
sexton_tile_names(roi)
}
\arguments{
\item{roi}{sf object, SpatVector, or numeric bbox}
}
\value{
Character vector of tile names in format "N00W060"
}
\description{
Generate Sexton tree cover tile names for a region
}
\keyword{internal}
