% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esacci_biomass.R
\name{validate_esacci_args}
\alias{validate_esacci_args}
\title{Validate ESA CCI Biomass year and version arguments}
\usage{
validate_esacci_args(year, version)
}
\arguments{
\item{year}{Numeric or character, year to validate}

\item{version}{Character, version to validate}
}
\value{
List with validated year and version
}
\description{
Validate ESA CCI Biomass year and version arguments
}
\keyword{internal}
