% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esacci_biomass.R
\name{validate_esacci_biomass_args}
\alias{validate_esacci_biomass_args}
\title{Validate ESA CCI Biomass year and version arguments}
\usage{
validate_esacci_biomass_args(esacci_biomass_year, esacci_biomass_version)
}
\arguments{
\item{esacci_biomass_year}{Numeric or character, year to validate}

\item{esacci_biomass_version}{Character, version to validate}
}
\value{
List with validated esacci_biomass_year and esacci_biomass_version
}
\description{
Validate ESA CCI Biomass year and version arguments
}
\keyword{internal}
