% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_extent}
\alias{validate_extent}
\title{Validate and standardize extent input}
\usage{
validate_extent(extent)
}
\arguments{
\item{extent}{Either an sf object or a numeric vector of length 4 (xmin, ymin, xmax, ymax)}
}
\value{
A numeric vector of length 4 with bbox coordinates
}
\description{
Validate and standardize extent input
}
\keyword{internal}
