% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{verify_download}
\alias{verify_download}
\title{Verify file download by comparing file sizes}
\usage{
verify_download(url, local_path)
}
\arguments{
\item{url}{Character, remote URL}

\item{local_path}{Character, local file path}
}
\value{
Logical, TRUE if sizes match, FALSE otherwise
}
\description{
Verify file download by comparing file sizes
}
\keyword{internal}
