% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esacci_biomass.R
\name{esacci_tile_names}
\alias{esacci_tile_names}
\title{Generate ESA CCI AGB tile names for a region}
\usage{
esacci_tile_names(
  roi,
  esacci_biomass_year,
  esacci_biomass_version,
  type = "agb"
)
}
\arguments{
\item{roi}{sf object, SpatVector, or numeric bbox}

\item{esacci_biomass_year}{Numeric, year}

\item{esacci_biomass_version}{Character, version string}

\item{type}{Character, either "agb" or "sd" for standard deviation}
}
\value{
Character vector of tile filenames
}
\description{
Generate ESA CCI AGB tile names for a region
}
\keyword{internal}
